backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-otp==1.2.2
cryptography==41.0.5
Pillow==10.1.0
python-magic==0.4.27
pyotp==2.9.0
bcrypt==4.0.1
django-cleanup==8.0.0
drf-yasg==1.21.7
python-dotenv==1.0.0
gunicorn==21.2.0
qrcode==7.4.2
setuptools>=65.5.1 